Prep for Spring 2019
Most of this is just cleaning and housekeeping. I fixed a few typos and tried
to clarify things in a few places. I updated the text so it's a bit more aware of
the fact that the students are turning things in via Canvas (or some other LMS)
now. I also replaced the user `peter` with the user `mcphee`, but I used
strikethroughs instead of deletion to make it easier for Peter to switch back
when he returns to the course.